k64ay/lesson_15/requirements.txt
protobuf==3.20.1
PyMySQL==1.0.2
python-dotenv==0.21.0
